pyproject.toml
[tool.poetry]
name = "ciplay-service"
version = "0.1.0"
description = ""
authors = ["subbotinav"]
readme = "README.md"
packages = [{include = "ciplay_service"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.88.0"}
asyncpg = "^0.27.0"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
sqlmodel = "^0.0.8"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
